Move FileReader() and IsJSONData() from x package to chunker.
